requirements.txt
streamlit
requests
pygsheets
python-dotenv
pyairtable
pytesseract
Pillow
numpy
